pyproject.toml
[project]
name = "enhanced-travel-agent"
version = "2.0.0"
description = "AI Travel Agent with LangChain, real-time APIs, and voice interface"
authors = [{name = "Your Name", email = "your.email@example.com"}]
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}

dependencies = [
    # Core AI/LangChain
    "openai>=1.0.0",
    "langchain>=0.1.0",
    "langchain-openai>=0.1.0",
    "langchain-community>=0.1.0",
    
    # Web framework
    "aiohttp>=3.8.0",
    "aiofiles>=23.0.0",
    
    # Data handling
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    
    # HTTP requests
    "httpx>=0.24.0",
    
    # Date handling
    "python-dateutil>=2.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 88
target-version = ['py38']

[tool.isort]
profile = "black"
line_length = 88

[project.scripts]
travel-agent = "main:main"